REFAC: Create useTextField as a reusable hook
